+++ remote-client/src/main/java/org/opensearch/remote/metadata/client/impl/RemoteClusterIndicesClient.java
@@ -151,7 +151,7 @@ public class RemoteClusterIndicesClient extends AbstractSdkClient {
                     .opType(request.overwriteIfExists() ? OpType.Index : OpType.Create)
                     .document(request.dataObject())
                     .tDocumentSerializer(new JsonTransformer.XContentObjectJsonpSerializer());
-                if (!Strings.isNullOrEmpty(request.id())) {
+                if (shouldUseId(request.id())) {
                     builder.id(request.id());
                 }
                 IndexRequest<?> indexRequest = builder.build();