+++ CHANGELOG.md
@@ -3,11 +3,11 @@ All notable changes to this project are documented in this file.
 
 Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
-## [Unreleased 3.2](https://github.com/opensearch-project/opensearch-remote-metadata-sdk/compare/3.1...HEAD)
+## [Unreleased 3.3](https://github.com/opensearch-project/opensearch-remote-metadata-sdk/compare/3.2...HEAD)
 ### Features
 ### Enhancements
 ### Bug Fixes
-- Avoid race condition putting the same document id in DDB Client ([#228](https://github.com/opensearch-project/opensearch-remote-metadata-sdk/pull/228))
+- Throw exception on empty string for put request ID ([#235](https://github.com/opensearch-project/opensearch-remote-metadata-sdk/pull/235))
 
 ### Infrastructure
 ### Documentation